When looking up subscribers using email. Use case insensitive check.
The casing of emails for users in your system may be different from
the casing used by users when they give their email address to Paddle
in order to perform a purchase.

This change makes the comparison of email addresses case insensitive.
The entire email address is treated as case insensitive which follows
the approach of django-allauth. As discussed here https://github.com/pennersr/django-allauth/issues/2233
that behaviour is technically incorrect as RFC 5321 says that the
local part of an email address should be treated as case sensitive
but it is not the approach usually taken.

If in the future, users of dj-paddle allow different users to sign up
to their site with with emails of the form user@email.com and
User@email.com then we could support this behaviour by adding an
extra flag to the settings.